pyproject.toml
[project]
name = "image_process_homework"
version = "0.1.0"
description = "A project for image processing homework"
authors = [
    { name="riemac", email="673944860@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scikit-image>=0.25.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
